fix xong

+++ finalModule/finalModule/src/main/java/com/codegym/finalModule/DTO/transaction/InventoryTransactionDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @Builder
 public class InventoryTransactionDTO {
 
-    private Integer id;
+    private Long id;
     @NotNull(message = "Giao dịch cần có thông tin sản phẩm !")
     private Integer product_id;
     @NotNull(message = "Không được để trống số lượng cần giao dịch !")
@@ -21,5 +21,4 @@ public class InventoryTransactionDTO {
     private Integer employee_id;
     private LocalDateTime created_at;
     private LocalDateTime updated_at;
-
 }
